[BUGFIX release] Ensure CP's are `_super` wrapped.
Prior to this change, a CP that was not included in the superclass would
not get re-wrapped. This means that when calling it from another function
that is required to be super wrapped would set that other function as
the `_super`.

This changes to ensure that all descriptors are wrapped if they include
`_super` (`hasSuper` check on platforms that support
`Function.prototype.toString`).